removed all course references
