--- newstack/new-auth-server/src/main/java/com/baeldung/lsso/LssoAuthorizationServerApp.java
@@ -1,38 +1,0 @@
-package com.baeldung.lsso;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.liquibase.LiquibaseAutoConfiguration;
-import org.springframework.boot.autoconfigure.web.ServerProperties;
-import org.springframework.boot.context.event.ApplicationReadyEvent;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.context.ApplicationListener;
-import org.springframework.context.annotation.Bean;
-
-import com.baeldung.lsso.config.KeycloakServerProperties;
-
-@SpringBootApplication(exclude = LiquibaseAutoConfiguration.class)
-@EnableConfigurationProperties(KeycloakServerProperties.class)
-public class LssoAuthorizationServerApp {
-
-    private static final Logger LOG = LoggerFactory.getLogger(LssoAuthorizationServerApp.class);
-
-    public static void main(String[] args) throws Exception {
-        SpringApplication.run(LssoAuthorizationServerApp.class, args);
-    }
-
-    @Bean
-    ApplicationListener<ApplicationReadyEvent> onApplicationReadyEventListener(ServerProperties serverProperties, KeycloakServerProperties keycloakServerProperties) {
-
-        return (evt) -> {
-
-            Integer port = serverProperties.getPort();
-            String keycloakContextPath = keycloakServerProperties.getContextPath();
-
-            LOG.info("Embedded Keycloak started: http://localhost:{}{} to use keycloak", port, keycloakContextPath);
-        };
-    }
-
-}
