+++ newstack/new-auth-server/src/main/java/com/baeldung/newstack/NewAuthorizationServerApp.java
@@ -1,0 +1,38 @@
+package com.baeldung.newstack;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.liquibase.LiquibaseAutoConfiguration;
+import org.springframework.boot.autoconfigure.web.ServerProperties;
+import org.springframework.boot.context.event.ApplicationReadyEvent;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.ApplicationListener;
+import org.springframework.context.annotation.Bean;
+
+import com.baeldung.newstack.config.KeycloakServerProperties;
+
+@SpringBootApplication(exclude = LiquibaseAutoConfiguration.class)
+@EnableConfigurationProperties(KeycloakServerProperties.class)
+public class NewAuthorizationServerApp {
+
+    private static final Logger LOG = LoggerFactory.getLogger(NewAuthorizationServerApp.class);
+
+    public static void main(String[] args) throws Exception {
+        SpringApplication.run(NewAuthorizationServerApp.class, args);
+    }
+
+    @Bean
+    ApplicationListener<ApplicationReadyEvent> onApplicationReadyEventListener(ServerProperties serverProperties, KeycloakServerProperties keycloakServerProperties) {
+
+        return (evt) -> {
+
+            Integer port = serverProperties.getPort();
+            String keycloakContextPath = keycloakServerProperties.getContextPath();
+
+            LOG.info("Embedded Keycloak started: http://localhost:{}{} to use keycloak", port, keycloakContextPath);
+        };
+    }
+
+}
