--- newstack/new-client/src/main/java/com/baeldung/lsso/LssoClientApplication.java
@@ -1,12 +1,0 @@
-package com.baeldung.lsso;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class LssoClientApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(LssoClientApplication.class, args);
-    }
-}
