--- newstack/new-client/src/main/java/com/baeldung/lsso/spring/SecurityConfig.java
@@ -1,35 +1,0 @@
-package com.baeldung.lsso.spring;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.oauth2.client.registration.ClientRegistrationRepository;
-import org.springframework.security.oauth2.client.web.OAuth2AuthorizedClientRepository;
-import org.springframework.security.oauth2.client.web.reactive.function.client.ServletOAuth2AuthorizedClientExchangeFilterFunction;
-import org.springframework.web.reactive.function.client.WebClient;
-
-@EnableWebSecurity
-public class SecurityConfig extends WebSecurityConfigurerAdapter {
-
-    @Override
-    protected void configure(HttpSecurity http) throws Exception {// @formatter:off
-        http.authorizeRequests()
-            .antMatchers("/").permitAll()
-            .anyRequest().authenticated()
-            .and()
-            .oauth2Login()
-            .and()
-            .logout().logoutSuccessUrl("/");
-    }// @formatter:on
-
-    @Bean
-    WebClient webClient(ClientRegistrationRepository clientRegistrationRepository, OAuth2AuthorizedClientRepository authorizedClientRepository) {
-        ServletOAuth2AuthorizedClientExchangeFilterFunction oauth2 = new ServletOAuth2AuthorizedClientExchangeFilterFunction(clientRegistrationRepository, authorizedClientRepository);
-        oauth2.setDefaultOAuth2AuthorizedClient(true);
-        return WebClient.builder()
-            .apply(oauth2.oauth2Configuration())
-            .build();
-    }
-
-}
