+++ newstack/new-client/src/main/java/com/baeldung/newstack/NewClientApplication.java
@@ -1,0 +1,12 @@
+package com.baeldung.newstack;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class NewClientApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(NewClientApplication.class, args);
+    }
+}
