+++ newstack/new-client/src/main/resources/application.yml
@@ -4,11 +4,11 @@ spring:
       client:
         registration:
           custom:
-            client-id: lssoClient
-            client-secret: lssoSecret
+            client-id: newClient
+            client-secret: newClientSecret
             scope: read,write
             authorization-grant-type: authorization_code
-            redirect-uri: http://localhost:8082/lsso-client/login/oauth2/code/custom
+            redirect-uri: http://localhost:8082/new-client/login/oauth2/code/custom
         provider:
           custom:
             authorization-uri: http://localhost:8083/auth/realms/baeldung/protocol/openid-connect/auth
@@ -21,7 +21,7 @@ spring:
 server: 
   port: 8082
   servlet: 
-    context-path: /lsso-client
+    context-path: /new-client
     
 logging: 
   level:
