+++ newstack/new-client/src/main/resources/templates/addproject.html
@@ -11,7 +11,7 @@
         <ul class="navbar-nav ml-auto">
             <li class="navbar-text">Hi, <span sec:authentication="name">username</span>&nbsp;&nbsp;&nbsp;
             </li>
-            <li class="nav-item"><a href="/lsso-client/logout" class="btn btn-outline-secondary">Logout</a></li>
+            <li class="nav-item"><a href="/new-client/logout" class="btn btn-outline-secondary">Logout</a></li>
         </ul>
     </nav>
