--- newstack/new-client/src/test/java/com/baeldung/lsso/ContextIntegrationTest.java
@@ -1,19 +1,0 @@
-package com.baeldung.lsso;
-
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-
-import com.baeldung.lsso.LssoClientApplication;
-
-@ExtendWith(SpringExtension.class)
-@SpringBootTest(classes = { LssoClientApplication.class })
-public class ContextIntegrationTest {
-
-    @Test
-    public void whenLoadApplication_thenSuccess() {
-
-    }
-
-}
