+++ newstack/new-resource-server/pom.xml
@@ -11,7 +11,7 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.2.3.RELEASE</version>
+        <version>2.2.2.RELEASE</version>
         <relativePath /> <!-- lookup parent from repository -->
     </parent>
