--- newstack/new-resource-server/src/main/java/com/baeldung/lsso/LssoResourceServerApp.java
@@ -1,13 +1,0 @@
-package com.baeldung.lsso;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class LssoResourceServerApp {
-
-    public static void main(String[] args) throws Exception {
-        SpringApplication.run(LssoResourceServerApp.class, args);
-    }
-
-}
