--- newstack/new-resource-server/src/main/java/com/baeldung/lsso/persistence/repository/IProjectRepository.java
@@ -1,8 +1,0 @@
-package com.baeldung.lsso.persistence.repository;
-
-import org.springframework.data.repository.PagingAndSortingRepository;
-
-import com.baeldung.lsso.persistence.model.Project;
-
-public interface IProjectRepository extends PagingAndSortingRepository<Project, Long> {
-}
