--- newstack/new-resource-server/src/main/java/com/baeldung/lsso/service/impl/ProjectServiceImpl.java
@@ -1,39 +1,0 @@
-package com.baeldung.lsso.service.impl;
-
-import java.time.LocalDate;
-import java.util.Optional;
-
-import org.springframework.stereotype.Service;
-import org.springframework.util.StringUtils;
-
-import com.baeldung.lsso.persistence.model.Project;
-import com.baeldung.lsso.persistence.repository.IProjectRepository;
-import com.baeldung.lsso.service.IProjectService;
-
-@Service
-public class ProjectServiceImpl implements IProjectService {
-
-    private IProjectRepository projectRepository;
-
-    public ProjectServiceImpl(IProjectRepository projectRepository) {
-        this.projectRepository = projectRepository;
-    }
-
-    @Override
-    public Optional<Project> findById(Long id) {
-        return projectRepository.findById(id);
-    }
-
-    @Override
-    public Project save(Project project) {
-        if (StringUtils.isEmpty(project.getId())) {
-            project.setDateCreated(LocalDate.now());
-        }
-        return projectRepository.save(project);
-    }
-
-    @Override
-    public Iterable<Project> findAll() {
-        return projectRepository.findAll();
-    }
-}
