--- newstack/new-resource-server/src/main/java/com/baeldung/lsso/web/controller/UserInfoController.java
@@ -1,18 +1,0 @@
-package com.baeldung.lsso.web.controller;
-
-import java.util.Collections;
-import java.util.Map;
-
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.oauth2.jwt.Jwt;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class UserInfoController {
-
-    @GetMapping("/user/info")
-    public Map<String, Object> getUserInfo(@AuthenticationPrincipal Jwt principal) {
-        return Collections.singletonMap("user_name", principal.getClaimAsString("preferred_username"));
-    }
-}
