+++ newstack/new-resource-server/src/main/java/com/baeldung/newstack/NewResourceServerApp.java
@@ -1,0 +1,13 @@
+package com.baeldung.newstack;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class NewResourceServerApp {
+
+    public static void main(String[] args) throws Exception {
+        SpringApplication.run(NewResourceServerApp.class, args);
+    }
+
+}
