+++ newstack/new-resource-server/src/main/java/com/baeldung/newstack/persistence/repository/IProjectRepository.java
@@ -1,0 +1,8 @@
+package com.baeldung.newstack.persistence.repository;
+
+import org.springframework.data.repository.PagingAndSortingRepository;
+
+import com.baeldung.newstack.persistence.model.Project;
+
+public interface IProjectRepository extends PagingAndSortingRepository<Project, Long> {
+}
