+++ newstack/new-resource-server/src/main/java/com/baeldung/newstack/spring/SecurityConfig.java
@@ -1,0 +1,25 @@
+package com.baeldung.newstack.spring;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.HttpMethod;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+
+@Configuration
+public class SecurityConfig extends WebSecurityConfigurerAdapter {
+
+    @Override
+    protected void configure(HttpSecurity http) throws Exception {// @formatter:off
+        http.authorizeRequests()
+              .antMatchers(HttpMethod.GET, "/user/info", "/api/projects/**")
+                .hasAuthority("SCOPE_read")
+              .antMatchers(HttpMethod.POST, "/api/projects")
+                .hasAuthority("SCOPE_write")
+              .anyRequest()
+                .authenticated()
+            .and()
+              .oauth2ResourceServer()
+                .jwt();
+    }//@formatter:on
+
+}
