+++ newstack/new-resource-server/src/main/java/com/baeldung/newstack/web/controller/ProjectController.java
@@ -1,0 +1,74 @@
+package com.baeldung.newstack.web.controller;
+
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.util.StringUtils;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
+
+import com.baeldung.newstack.persistence.model.Project;
+import com.baeldung.newstack.service.IProjectService;
+import com.baeldung.newstack.web.dto.ProjectDto;
+
+@RestController
+@RequestMapping(value = "/api/projects")
+public class ProjectController {
+
+    private IProjectService projectService;
+
+    public ProjectController(IProjectService projectService) {
+        this.projectService = projectService;
+    }
+
+    @GetMapping(value = "/{id}")
+    public ProjectDto findOne(@PathVariable Long id) {
+        Project entity = projectService.findById(id)
+            .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+        return convertToDto(entity);
+    }
+
+    @ResponseStatus(HttpStatus.CREATED)
+    @PostMapping
+    public void create(@RequestBody ProjectDto newProject) {
+        Project entity = convertToEntity(newProject);
+        this.projectService.save(entity);
+    }
+
+    @GetMapping
+    public Collection<ProjectDto> findAll() {
+        Iterable<Project> projects = this.projectService.findAll();
+        List<ProjectDto> projectDtos = new ArrayList<>();
+        projects.forEach(p -> projectDtos.add(convertToDto(p)));
+        return projectDtos;
+    }
+
+    @PutMapping("/{id}")
+    public ProjectDto updateProject(@PathVariable("id") Long id, @RequestBody ProjectDto updatedProject) {
+        Project projectEntity = convertToEntity(updatedProject);
+        return this.convertToDto(this.projectService.save(projectEntity));
+    }
+
+    protected ProjectDto convertToDto(Project entity) {
+        ProjectDto dto = new ProjectDto(entity.getId(), entity.getName(), entity.getDateCreated());
+
+        return dto;
+    }
+
+    protected Project convertToEntity(ProjectDto dto) {
+        Project project = new Project(dto.getName(), dto.getDateCreated());
+        if (!StringUtils.isEmpty(dto.getId())) {
+            project.setId(dto.getId());
+        }
+        return project;
+    }
+}
