+++ newstack/new-resource-server/src/main/java/com/baeldung/newstack/web/dto/ProjectDto.java
@@ -1,0 +1,49 @@
+package com.baeldung.newstack.web.dto;
+
+import java.time.LocalDate;
+
+import org.springframework.format.annotation.DateTimeFormat;
+import org.springframework.format.annotation.DateTimeFormat.ISO;
+
+public class ProjectDto {
+
+    private Long id;
+
+    private String name;
+
+    @DateTimeFormat(iso = ISO.DATE)
+    private LocalDate dateCreated;
+
+    public ProjectDto() {
+    }
+
+    public ProjectDto(Long id, String name, LocalDate dateCreated) {
+        this.id = id;
+        this.name = name;
+        this.dateCreated = dateCreated;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public LocalDate getDateCreated() {
+        return dateCreated;
+    }
+
+    public void setDateCreated(LocalDate dateCreated) {
+        this.dateCreated = dateCreated;
+    }
+}
