+++ newstack/new-resource-server/src/main/resources/application.yml
@@ -1,7 +1,7 @@
 server: 
   port: 8081
   servlet: 
-    context-path: /lsso-resource-server
+    context-path: /new-resource-server
 
 ####### resource server configuration properties
 spring:
